requirements.txt
loguru
colorama
tqdm

numpy==1.26.4
scikit-learn
pandas
jupyter
ipywidgets==7.7.1
openpyxl
einops
matplotlib
plotly
mne
seaborn
pyyaml
torch==2.1.0
torchvision==0.16
torchaudio==2.1.0
torchmetrics
lightning==2.4.0
transformers==4.45.0
wandb




